[US75088] Use a SHA1 hash on the advisory lock
Since sha1 will consistently produce the same value, it's functionally
as good as the original shard id value for getting the advisory lock. On
mysql 5.7, the advisory lock can only be 64 characters long. SHA1 values
are 40 characters long we can use that value to stay within the limit.

Technically this creates a small risk of disordering messages during
deployment, but not if the workers are paused during the deployment,
which is recommended if upgrading a conductor to the latest version.